requirements.txt
Flask==3.0.2
requests==2.25.1
beautifulsoup4==4.9.3
pyyaml==6.0.1
openai==1.14.1